CB-15274 generalize freeipa instance await and cloudbreak instance await code
