+++ integration-test/src/main/java/com/sequenceiq/it/cloudbreak/context/TestContext.java
@@ -29,7 +29,6 @@ import org.springframework.context.ApplicationContextAware;
 import org.testng.ITestResult;
 import org.testng.Reporter;
 
-import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.base.InstanceStatus;
 import com.sequenceiq.cloudbreak.auth.crn.Crn;
 import com.sequenceiq.cloudbreak.common.mappable.CloudPlatform;
 import com.sequenceiq.it.TestParameter;
@@ -49,7 +48,6 @@ import com.sequenceiq.it.cloudbreak.cloud.v4.CloudProviderProxy;
 import com.sequenceiq.it.cloudbreak.cloud.v4.CommonCloudProperties;
 import com.sequenceiq.it.cloudbreak.cloud.v4.CommonClusterManagerProperties;
 import com.sequenceiq.it.cloudbreak.dto.CloudbreakTestDto;
-import com.sequenceiq.it.cloudbreak.dto.freeipa.FreeIpaTestDto;
 import com.sequenceiq.it.cloudbreak.exception.TestFailException;
 import com.sequenceiq.it.cloudbreak.finder.Attribute;
 import com.sequenceiq.it.cloudbreak.finder.Capture;
@@ -59,7 +57,6 @@ import com.sequenceiq.it.cloudbreak.util.ErrorLogMessageProvider;
 import com.sequenceiq.it.cloudbreak.util.ResponseUtil;
 import com.sequenceiq.it.cloudbreak.util.wait.FlowUtil;
 import com.sequenceiq.it.cloudbreak.util.wait.service.ResourceAwait;
-import com.sequenceiq.it.cloudbreak.util.wait.service.freeipa.instance.FreeIpaInstanceAwait;
 import com.sequenceiq.it.cloudbreak.util.wait.service.instance.InstanceAwait;
 
 import io.opentracing.Tracer;
@@ -135,9 +132,6 @@ public abstract class TestContext implements ApplicationContextAware {
 
     @Inject
     private InstanceAwait instanceAwait;
-
-    @Inject
-    private FreeIpaInstanceAwait freeIpaInstanceAwait;
 
     @Inject
     private ErrorLogMessageProvider errorLogMessageProvider;
@@ -979,11 +973,13 @@ public abstract class TestContext implements ApplicationContextAware {
         return entity;
     }
 
-    public <T extends CloudbreakTestDto> T awaitForInstance(T entity, Map<List<String>, InstanceStatus> desiredStatuses, RunningParameter runningParameter) {
+    public <T extends CloudbreakTestDto, E extends Enum<E>> T awaitForInstance(T entity, Map<List<String>, E> desiredStatuses,
+            RunningParameter runningParameter) {
         return awaitForInstance(entity, desiredStatuses, runningParameter, getPollingDurationInMills());
     }
 
-    public <T extends CloudbreakTestDto> T awaitForInstance(T entity, Map<List<String>, InstanceStatus> desiredStatuses, RunningParameter runningParameter,
+    public <T extends CloudbreakTestDto, E extends Enum<E>> T awaitForInstance(T entity, Map<List<String>, E> desiredStatuses,
+            RunningParameter runningParameter,
             Duration pollingInterval) {
         checkShutdown();
         if (!getExceptionMap().isEmpty() && runningParameter.isSkipOnFail()) {
@@ -994,27 +990,6 @@ public abstract class TestContext implements ApplicationContextAware {
         String key = getKeyForAwait(entity, entity.getClass(), runningParameter);
         CloudbreakTestDto awaitEntity = get(key);
         instanceAwait.await(awaitEntity, desiredStatuses, getTestContext(), runningParameter, pollingInterval, maxRetry);
-        return entity;
-    }
-
-    public <T extends CloudbreakTestDto> T awaitForFreeIpaInstance(T entity, Map<List<String>,
-            com.sequenceiq.freeipa.api.v1.freeipa.stack.model.common.instance.InstanceStatus> desiredStatuses,
-            RunningParameter runningParameter) {
-        return awaitForFreeIpaInstance(entity, desiredStatuses, runningParameter, getPollingDurationInMills());
-    }
-
-    public <T extends CloudbreakTestDto> T awaitForFreeIpaInstance(T entity,
-            Map<List<String>, com.sequenceiq.freeipa.api.v1.freeipa.stack.model.common.instance.InstanceStatus> desiredStatuses,
-            RunningParameter runningParameter, Duration pollingInterval) {
-        checkShutdown();
-        if (!getExceptionMap().isEmpty() && runningParameter.isSkipOnFail()) {
-            Log.await(LOGGER, String.format("Cloudbreak await for instance should be skipped because of previous error. awaitforinstance [%s]",
-                    desiredStatuses));
-            return entity;
-        }
-        String key = getKeyForAwait(entity, entity.getClass(), runningParameter);
-        FreeIpaTestDto awaitEntity = get(key);
-        freeIpaInstanceAwait.await(awaitEntity, desiredStatuses, getTestContext(), runningParameter, pollingInterval, maxRetry);
         return entity;
     }
