+++ integration-test/src/main/java/com/sequenceiq/it/cloudbreak/testcase/mock/FreeIpaStartStopTest.java
@@ -46,8 +46,8 @@ public class FreeIpaStartStopTest extends AbstractMockTest {
                 .when(freeIpaTestClient.stop())
                 .await(Status.STOPPED)
                 .when(freeIpaTestClient.start())
-                .awaitForHealthyInstances()
                 .await(Status.AVAILABLE)
+                .awaitForHealthyInstances()
                 .when(freeIpaTestClient.delete())
                 .await(Status.DELETE_COMPLETED)
                 .when(freeIpaAuditGrpcServiceAssertion::create)
